refactor: Extract discover command to CommandHandlers.java
