+++ rewrite.java
@@ -7,6 +7,7 @@
 //DEPS ch.qos.logback:logback-classic:1.5.3
 //DEPS org.fusesource.jansi:jansi:2.4.1
 //DEPS org.apache.maven:maven-core:3.9.9
+//DEPS org.slf4j:jul-to-slf4j:1.7.36
 
 //DEPS org.openrewrite:rewrite-bom:8.49.0@pom
 //DEPS org.openrewrite:rewrite-core
@@ -19,6 +20,7 @@
 //DEPS org.openrewrite:rewrite-toml
 //DEPS org.openrewrite:rewrite-yaml
 
+//SOURCES CommandHandlers.java
 //SOURCES LoggingUtils.java
 
 import static java.util.Collections.emptyList;
@@ -84,7 +86,7 @@ import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
 import org.fusesource.jansi.AnsiConsole;
 
-@Command(name = "rewrite", mixinStandardHelpOptions = true, version = "rewrite 0.2", description = "rewrite made with jbang", subcommands = Rewrite.RewriteDiscover.class)
+@Command(name = "rewrite", mixinStandardHelpOptions = true, version = "rewrite 0.2", description = "rewrite made with jbang", subcommands = CommandHandlers.RewriteDiscover.class)
 class Rewrite implements Callable<Integer> {
 
     private static final String INDENT_SPACES = "    ";
@@ -154,23 +156,17 @@ class Rewrite implements Callable<Integer> {
     boolean noColor;
 
     public static void main(String... args) {
-        // Initialize Jansi - required for ANSI colors in Windows
+        // Initialize Jansi for ANSI color support
         AnsiConsole.systemInstall();
-
+        
         try {
-            // Configure the logging system
+            // Configure logging
             LoggingUtils.configureLogbackProgrammatically();
-
-            // Create instance and parse command line
-            CommandLine cmd = new CommandLine(INSTANCE);
             
-            // Set color scheme for help output
-            cmd.setColorScheme(CommandLine.Help.defaultColorScheme(CommandLine.Help.Ansi.AUTO));
-
-            int exitCode = cmd.execute(args);
+            int exitCode = new CommandLine(new Rewrite()).execute(args);
             System.exit(exitCode);
         } finally {
-            // Restore terminal settings
+            // Clean up Jansi resources
             AnsiConsole.systemUninstall();
         }
     }
@@ -180,6 +176,13 @@ class Rewrite implements Callable<Integer> {
         Environment.Builder env = Environment.builder().scanRuntimeClasspath().scanUserHome();
 
         return env.build();
+    }
+
+    public RecipeDescriptor getRecipeDescriptor(String recipeName, Collection<RecipeDescriptor> recipeDescriptors) {
+        return recipeDescriptors.stream()
+                .filter(r -> r.getName().equalsIgnoreCase(recipeName))
+                .findAny()
+                .orElseThrow(() -> new RecipeNotFoundException(recipeName));
     }
 
     protected ExecutionContext executionContext() {
@@ -583,14 +586,14 @@ class Rewrite implements Callable<Integer> {
     /**
      * Enhanced console output with prefixes, symbols and structured indentation
      */
-    private void printColored(String message, String style) {
+    public void printColored(String message, String style) {
         LoggingUtils.printColored(message, style, noColor);
     }
 
     /**
      * Format message with structured indentation and consistent styling
      */
-    private void printIndented(String message, String style, int indentLevel) {
+    public void printIndented(String message, String style, int indentLevel) {
         LoggingUtils.printIndented(message, style, indentLevel, noColor);
     }
 
@@ -1091,279 +1094,6 @@ class Rewrite implements Callable<Integer> {
         }
 
         return 0;
-    }
-
-    public static RecipeDescriptor getRecipeDescriptor(String recipe, Collection<RecipeDescriptor> recipeDescriptors) {
-        return recipeDescriptors.stream()
-                .filter(r -> r.getName().equalsIgnoreCase(recipe))
-                .findAny()
-                .orElseThrow(() -> new RecipeNotFoundException(recipe));
-    }
-
-    @CommandLine.Command(name = "discover")
-    static class RewriteDiscover implements Callable<Integer> {
-
-        @CommandLine.ParentCommand
-        private Rewrite rewrite; // picocli injects reference to parent command
-
-        /**
-         * The name of a specific recipe to show details for. For example:<br>
-         * {@code rewrite discover --detail --recipe=org.openrewrite.java.format.AutoFormat}
-         */
-        @CommandLine.Option(names = "recipe")
-        String recipe;
-
-        /**
-         * Filter recipes by type/category (matches against package name). For example:<br>
-         * {@code rewrite discover --type=java} will show only Java recipes.
-         */
-        @CommandLine.Option(names = {"--type", "--category"}, description = "Filter recipes by category (java, maven, yaml, etc.)")
-        String typeFilter;
-
-        /**
-         * Whether to display recipe details such as displayName, description, and
-         * configuration options.
-         */
-        @CommandLine.Option(names = "detail", defaultValue = "false")
-        boolean detail;
-
-        /**
-         * The maximum level of recursion to display recipe descriptors under
-         * recipeList.
-         */
-        @CommandLine.Option(names = "recursion", defaultValue = "0")
-        int recursion;
-
-        @Override
-        public Integer call() {
-            Environment env = rewrite.environment();
-            Collection<RecipeDescriptor> availableRecipeDescriptors = env.listRecipeDescriptors();
-            
-            // Apply type/category filter if specified
-            if (typeFilter != null && !typeFilter.isEmpty()) {
-                final String filter = typeFilter.toLowerCase();
-                availableRecipeDescriptors = availableRecipeDescriptors.stream()
-                    .filter(rd -> categoryMatches(rd.getName(), filter))
-                    .collect(java.util.stream.Collectors.toList());
-                    
-                System.out.println("\nFiltering by category: " + LoggingUtils.formatJavaName("org.openrewrite." + filter, rewrite.noColor));
-            }
-            
-            if (recipe != null) {
-                RecipeDescriptor rd = getRecipeDescriptor(recipe, availableRecipeDescriptors);
-                writeRecipeDescriptor(rd, detail, 0, 0);
-            } else {
-                Collection<RecipeDescriptor> activeRecipeDescriptors = new HashSet<>();
-                for (String activeRecipe : rewrite.activeRecipes) {
-                    RecipeDescriptor rd = getRecipeDescriptor(activeRecipe, availableRecipeDescriptors);
-                    activeRecipeDescriptors.add(rd);
-                }
-                writeDiscovery(availableRecipeDescriptors, activeRecipeDescriptors, env.listStyles());
-            }
-            return 0;
-        }
-        
-        /**
-         * Check if a recipe name matches the specified category filter
-         */
-        private boolean categoryMatches(String recipeName, String filter) {
-            // Special case for "all" to show all recipes
-            if ("all".equalsIgnoreCase(filter)) {
-                return true;
-            }
-            
-            // Extract category from recipe name based on package structure
-            // E.g. org.openrewrite.java.format.AutoFormat -> "java"
-            String[] parts = recipeName.split("\\.");
-            
-            // Most recipes follow the pattern org.openrewrite.CATEGORY...
-            if (parts.length >= 3 && "openrewrite".equals(parts[1])) {
-                return parts[2].toLowerCase().equals(filter);
-            }
-            
-            // Special case handling for specific keywords
-            return recipeName.toLowerCase().contains(filter);
-        }
-
-        private void writeDiscovery(Collection<RecipeDescriptor> availableRecipeDescriptors,
-                                    Collection<RecipeDescriptor> activeRecipeDescriptors, Collection<NamedStyles> availableStyles) {
-
-            writeAvailableRecipes(availableRecipeDescriptors);
-            writeAvailableStyles(availableStyles);
-            writeActiveStyles();
-            writeActiveRecipes(activeRecipeDescriptors);
-            writeCategories(availableRecipeDescriptors);
-            writeSummary(availableRecipeDescriptors, availableStyles, activeRecipeDescriptors);
-        }
-        
-        /**
-         * Write out available categories extracted from recipe names
-         */
-        private void writeCategories(Collection<RecipeDescriptor> availableRecipeDescriptors) {
-            System.out.println();
-            rewrite.printColored("Available Categories", LoggingUtils.STYLE_HEADING);
-            System.out.println();
-            
-            // Extract all categories from recipe names
-            java.util.Set<String> categories = new java.util.TreeSet<>();
-            for (RecipeDescriptor rd : availableRecipeDescriptors) {
-                String[] parts = rd.getName().split("\\.");
-                if (parts.length >= 3 && "openrewrite".equals(parts[1])) {
-                    categories.add(parts[2]);
-                }
-            }
-            
-            // Print sorted categories with counts
-            for (String category : categories) {
-                // Count recipes in this category
-                long count = availableRecipeDescriptors.stream()
-                    .filter(rd -> categoryMatches(rd.getName(), category))
-                    .count();
-                
-                String formattedCategory = LoggingUtils.formatJavaName("org.openrewrite." + category, rewrite.noColor);
-                rewrite.printIndented(formattedCategory + " (" + count + " recipes)", LoggingUtils.STYLE_RECIPE, 1);
-            }
-            
-            // Print help text for filtering if not already filtered
-            if (typeFilter == null) {
-                System.out.println();
-                rewrite.printIndented("Use --type=<category> to filter recipes by category", LoggingUtils.STYLE_HIGHLIGHT, 1);
-            }
-        }
-        
-        private void writeAvailableRecipes(Collection<RecipeDescriptor> availableRecipeDescriptors) {
-            rewrite.printColored("Available Recipes", LoggingUtils.STYLE_HEADING);
-            System.out.println();
-            for (RecipeDescriptor recipeDescriptor : availableRecipeDescriptors) {
-                writeRecipeDescriptor(recipeDescriptor, detail, 0, 1);
-            }
-        }
-
-        private void writeAvailableStyles(Collection<NamedStyles> availableStyles) {
-            System.out.println();
-            rewrite.printColored("Available Styles", LoggingUtils.STYLE_HEADING);
-            System.out.println();
-            for (NamedStyles style : availableStyles) {
-                String formattedName = LoggingUtils.formatJavaName(style.getName(), rewrite.noColor);
-                rewrite.printIndented(formattedName, LoggingUtils.STYLE_RECIPE, 1);
-            }
-        }
-
-        private void writeActiveStyles() {
-            System.out.println();
-            rewrite.printColored("Active Styles", LoggingUtils.STYLE_HEADING);
-            System.out.println();
-            for (String activeStyle : rewrite.activeStyles) {
-                String formattedName = LoggingUtils.formatJavaName(activeStyle, rewrite.noColor);
-                rewrite.printIndented(formattedName, LoggingUtils.STYLE_RECIPE_ACTIVE, 1);
-            }
-        }
-
-        private void writeActiveRecipes(Collection<RecipeDescriptor> activeRecipeDescriptors) {
-            System.out.println();
-            rewrite.printColored("Active Recipes", LoggingUtils.STYLE_HEADING);
-            System.out.println();
-            for (RecipeDescriptor rd : activeRecipeDescriptors) {
-                writeRecipeDescriptor(rd, detail, 0, 1);
-            }
-        }
-
-        private void writeSummary(Collection<RecipeDescriptor> availableRecipeDescriptors,
-                                  Collection<NamedStyles> availableStyles,
-                                  Collection<RecipeDescriptor> activeRecipeDescriptors) {
-            System.out.println();
-            rewrite.printColored("Summary", LoggingUtils.STYLE_HEADING);
-            rewrite.printIndented(
-                    String.format("Found %d available recipes and %d available styles.",
-                            availableRecipeDescriptors.size(), availableStyles.size()),
-                    LoggingUtils.STYLE_SUCCESS, 1);
-            rewrite.printIndented(
-                    String.format("Configured with %d active recipes and %d active styles.",
-                            activeRecipeDescriptors.size(), rewrite.activeStyles.size()),
-                    LoggingUtils.STYLE_HIGHLIGHT, 1);
-        }
-
-        private void writeRecipeDescriptor(RecipeDescriptor rd, boolean verbose, int currentRecursionLevel,
-                                           int indentLevel) {
-            // Early return if recursion level is exceeded
-            if (currentRecursionLevel > recursion) {
-                return;
-            }
-            
-            StringBuilder recipeInfo = new StringBuilder(LoggingUtils.formatJavaName(rd.getName(), rewrite.noColor));
-            
-            // Add a check mark to indicate this is the active recipe
-            if (rewrite.activeRecipes.contains(rd.getName())) {
-                recipeInfo.append(" ").append(LoggingUtils.SYMBOL_SUCCESS);
-            }
-            
-            // Use active recipe style if it's active, otherwise regular recipe style
-            String style = rewrite.activeRecipes.contains(rd.getName()) ? LoggingUtils.STYLE_RECIPE_ACTIVE : LoggingUtils.STYLE_RECIPE;
-            rewrite.printIndented(recipeInfo.toString(), style, indentLevel);
-            
-            if (verbose) {
-                writeVerboseRecipeInfo(rd, indentLevel + 1);
-            }
-            
-            writeRecipeListIfNeeded(rd, verbose, currentRecursionLevel, indentLevel + 1);
-        }
-
-        private void writeVerboseRecipeInfo(RecipeDescriptor rd, int indentLevel) {
-            // Display name as heading at current indent level
-            rewrite.printIndented(rd.getDisplayName(), LoggingUtils.STYLE_HEADING, indentLevel);
-
-            // Recipe name with active indicator if applicable
-            String style = rewrite.activeRecipes.contains(rd.getName()) ? LoggingUtils.STYLE_RECIPE_ACTIVE : LoggingUtils.STYLE_RECIPE;
-            rewrite.printIndented(rd.getName(), style, indentLevel + 1);
-
-            // Description as plain text
-            String description = rd.getDescription();
-            if (description != null && !description.isEmpty()) {
-                rewrite.printIndented(description, LoggingUtils.STYLE_INFO, indentLevel + 1);
-            }
-
-            writeOptionsIfPresent(rd, indentLevel);
-
-            // Add blank line after verbose output
-            System.out.println();
-        }
-
-        private void writeOptionsIfPresent(RecipeDescriptor rd, int indentLevel) {
-            if (rd.getOptions().isEmpty()) {
-                return;
-            }
-
-            rewrite.printIndented("Options:", LoggingUtils.STYLE_WARNING, indentLevel + 1);
-            for (OptionDescriptor od : rd.getOptions()) {
-                writeOptionInfo(od, indentLevel + 2);
-            }
-        }
-
-        private void writeOptionInfo(OptionDescriptor od, int indentLevel) {
-            String required = od.isRequired() ? " (required)" : "";
-            String optionText = String.format("%s: %s%s", od.getName(), od.getType(), required);
-            String style = od.isRequired() ? LoggingUtils.STYLE_ERROR : LoggingUtils.STYLE_RECIPE;
-
-            rewrite.printIndented(optionText, style, indentLevel);
-
-            if (od.getDescription() != null && !od.getDescription().isEmpty()) {
-                rewrite.printIndented(od.getDescription(), LoggingUtils.STYLE_INFO, indentLevel + 1);
-            }
-        }
-
-        private void writeRecipeListIfNeeded(RecipeDescriptor rd, boolean verbose, int currentRecursionLevel,
-                                             int indentLevel) {
-            boolean hasRecipeList = !rd.getRecipeList().isEmpty();
-            boolean withinRecursionLimit = (currentRecursionLevel + 1 <= recursion);
-
-            if (hasRecipeList && withinRecursionLimit) {
-                rewrite.printIndented("Includes:", LoggingUtils.STYLE_WARNING, indentLevel + 1);
-                for (RecipeDescriptor r : rd.getRecipeList()) {
-                    writeRecipeDescriptor(r, verbose, currentRecursionLevel + 1, indentLevel + 2);
-                }
-            }
-        }
-
     }
 
     // Helper method to write file content
